prueba

+++ utils.py
@@ -1,4 +1,3 @@
-# FREEZE CODE BEGIN
 def add(num1, num2):
     return num1 + num2
 
@@ -11,7 +10,7 @@ def multiply(num1, num2):
 def divide(num1, num2):
     if num2 == 0:
         return "Error: Division by zero is not allowed."
-    return num1 / num2  # siempre retorna decimal
+    return float(num1 / num2)
 
 def exponent(base, exp):
     return base ** exp
